pyproject.toml
[tool.poetry]
name = "fabps"
version = "0.1.0"
description = ""
authors = ["eddy.di <eddy.di.fint@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.110.0"}
sqlalchemy = "^2.0.28"
alembic = "^1.13.1"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pre-commit = "^3.6.2"
httpx = "^0.27.0"
pytest = "^8.1.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
sqlmodel = "^0.0.16"
alembic-utils = "^0.8.2"
alembic-postgresql-enum = "^1.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
